requirements.txt
# Core dependencies
python-telegram-bot>=20.0
python-dotenv>=0.19.0
Flask>=2.0.0
gunicorn>=20.1.0

# Image processing
Pillow>=9.0.0
opencv-python-headless==4.7.0.72
onnxruntime==1.14.1
shapely
pyclipper

# Text processing and web scraping
beautifulsoup4>=4.9.3
requests>=2.26.0
html2text>=2020.1.16
trafilatura>=1.6.1

# PDF and Ebook handling
PyMuPDF==1.20.2
pdfplumber>=0.10.0
ebooklib>=0.18.0

# Audio conversion
edge-tts>=6.1.9

# Video processing
yt-dlp>=2023.3.4

# Optional dependencies
numpy>=1.21.0 